Add Tubestuff's Automatic Crafting Table Mk2 as an importable tile entity in crafting logistics pipes (#5)
Let's you press the import button on Crafting Logistics Pipes if a valid crafting recipe is found in an attached Autocrafting Table Mk2 from Tubestuff.

The reason for this inclusion is that both BC's Autocrafting Table and Xycraft's Fabricator aren't ideal in a wall of autocrafting tables. BC tables need to have the actual item in them and don't work on a ghost item. Xycraft Fabricators steal inputs from adjacent Fabricators, which is also not ideal for the usecase of LP autocrafting.

Tubestuff tables allow recipes to be inserted from NEI and don't steal inputs from other crafting tables, making them ideal for a "compact" wall of LP autocrafting.

Reviewed-on: https://git.sleeping.town/Rewind/UpsilonFixes/pulls/5
Co-authored-by: ThePixelbrain <thepixelbrain@noreply.git.sleeping.town>
Co-committed-by: ThePixelbrain <thepixelbrain@noreply.git.sleeping.town>