Fix DartCraft mob spawn registration (#6)
DartCraft's Ender Tots spawn in places where they should not.

It spawns **a lot** in the Twilight Forest. While I haven't looked at the TF code, mobs in the dim seem to not follow vanilla spawning rules, as mobs don't spawn on the surface even in dark areas. The only mob that does is the Ender Tot, resulting in massive spawn rates.
Moving the mob spawn registration from post init to init fixes this. Other mods like Thaumcraft do their registration in init as well, though I don't know if this is a proper convention.

Ender Tots also spawn in Mushroom Island Biomes when it's dark. I added a manual exception for this biome, as I don't think there is any proper rule to check for.

Co-authored-by: ThePixelbrain <19214217+ThePixelbrain@users.noreply.github.com>
Reviewed-on: https://git.sleeping.town/Rewind/UpsilonFixes/pulls/6
Co-authored-by: ThePixelbrain <thepixelbrain@noreply.git.sleeping.town>
Co-committed-by: ThePixelbrain <thepixelbrain@noreply.git.sleeping.town>