Fix ChickenChunks awayTimeout (#9)
Fixes ChickenChunks awayTimeout feature.

If enabled in the config it should keep chunkloaders running for a specified time after the player logged off.
This feature never worked, deactivating all chunk loaders of offline players instantly unless the `awayTimeout` is set to `0`, which is interpreted as infinite loading and also the default value of the config.

The feature breaks because the required calls in the `load` method of the `ChunkLoaderManager` class are in the wrong order.

The method tries to generate tickets for chunkloaders before checking the eligibility of the player. Therefore if the `awayTimeout` is set to anything but `0` in the config file, the Chunk Loaders stop working for offline players even if they should still be active.
This bytecode patch reorders the method calls, checking the eligibility first and then generating tickets for all eligible chunkloaders.

Co-authored-by: ThePixelbrain <19214217+ThePixelbrain@users.noreply.github.com>
Reviewed-on: https://git.sleeping.town/Rewind/UpsilonFixes/pulls/9
Co-authored-by: ThePixelbrain <thepixelbrain@noreply.git.sleeping.town>
Co-committed-by: ThePixelbrain <thepixelbrain@noreply.git.sleeping.town>